chore(worker-service): rename usecases folder
