+++ apps/worker-service/src/app/workflow/usecases/update-job-status/set-job-as-failed.usecase.ts
