+++ apps/worker-service/src/app/workflow/usecases/update-job-status/update-job-status.command.ts
