+++ apps/worker-service/src/app/workflow/workflow.module.ts
@@ -44,7 +44,7 @@ import {
   SetJobAsFailed,
   UpdateJobStatus,
   WebhookFilterBackoffStrategy,
-} from './usecases2';
+} from './usecases';
 
 const USE_CASES = [
   AddJob,
